backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.12.0
python-dotenv==1.0.0
numpy==1.26.3
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

